refactor: add voters table & top accounts mobile styling (#319)
