+++ app/Http/Livewire/WalletVoterTable.php
@@ -4,34 +4,56 @@ declare(strict_types=1);
 
 namespace App\Http\Livewire;
 
+use App\Http\Livewire\Concerns\HasTablePagination;
 use App\Models\Scopes\OrderByBalanceScope;
 use App\Models\Wallet;
 use App\ViewModels\ViewModelFactory;
-use ARKEcosystem\Foundation\UserInterface\Http\Livewire\Concerns\HasPagination;
+use App\ViewModels\WalletViewModel;
+use Illuminate\Pagination\LengthAwarePaginator;
 use Illuminate\View\View;
 use Livewire\Component;
 
+/** @property LengthAwarePaginator $wallets */
 final class WalletVoterTable extends Component
 {
-    use HasPagination;
+    use HasTablePagination;
 
-    public string $publicKey;
+    public const PER_PAGE = 10;
 
-    public string $username;
+    public ?string $publicKey;
 
     /** @var mixed */
     protected $listeners = ['currencyChanged' => '$refresh'];
 
-    public function mount(string $publicKey, string $username): void
+    public function mount(WalletViewModel $wallet): void
     {
-        $this->publicKey = $publicKey;
-        $this->username  = $username;
+        $this->publicKey = $wallet->publicKey();
     }
 
     public function render(): View
     {
         return view('livewire.wallet-voter-table', [
-            'wallets' => ViewModelFactory::paginate(Wallet::where('attributes->vote', $this->publicKey)->withScope(OrderByBalanceScope::class)->paginate()),
+            'wallets' => ViewModelFactory::paginate($this->wallets),
         ]);
+    }
+
+    public function getNoResultsMessageProperty(): ?string
+    {
+        if ($this->wallets->total() === 0) {
+            return trans('tables.wallets.no_results');
+        }
+
+        return null;
+    }
+
+    public function getWalletsProperty(): LengthAwarePaginator
+    {
+        if ($this->publicKey === null) {
+            return new LengthAwarePaginator([], 0, $this->perPage);
+        }
+
+        return Wallet::where('attributes->vote', $this->publicKey)
+            ->withScope(OrderByBalanceScope::class)
+            ->paginate($this->perPage);
     }
 }
