+++ composer.json
@@ -12,7 +12,7 @@
         "ext-gmp": "*",
         "ardenthq/arkvault-url": "^1.1",
         "arkecosystem/crypto": "^1.8",
-        "arkecosystem/foundation": "^13.5",
+        "arkecosystem/foundation": "^13.6",
         "blade-ui-kit/blade-icons": "^1.5",
         "brick/math": "^0.11",
         "doctrine/dbal": "^3.0",
