+++ public/mix-manifest.json
@@ -2,7 +2,7 @@
     "/js/app.js": "/js/app.js?id=dad09d0514f44cb93235949966e05322",
     "/js/clipboard.js": "/js/clipboard.js?id=80d900f861e2aacb0e23e914068d0de9",
     "/js/manifest.js": "/js/manifest.js?id=d7f05315a876420076a2066217d34e8f",
-    "/css/app.css": "/css/app.css?id=fb8b7d1fbb08f8b563f841cd8579c3c1",
+    "/css/app.css": "/css/app.css?id=c87063e5ee74579a4039829f0ca99217",
     "/js/vendor.js": "/js/vendor.js?id=a0fbceff2e2c3da1c0b24a701d9c2403",
     "/images/errors/401.svg": "/images/errors/401.svg?id=f0752ffe6436f6cb98210be88dd36eb5",
     "/images/errors/401_dark.svg": "/images/errors/401_dark.svg?id=ade1bb3ec92d4001f1af918885b2b5c7",
