+++ resources/lang/en/tables.php
@@ -22,6 +22,11 @@ return [
         ],
     ],
 
+    'wallets' => [
+        'balance_currency' => 'Balance (:currency)',
+        'no_results'       => 'This delegate does not currently have any voters.',
+    ],
+
     'exchanges' => [
         'price'  => 'Price (:currency)',
         'volume' => 'Volume (:currency)',
