+++ resources/views/components/tables/desktop/skeleton/wallet-transactions.blade.php
@@ -1,3 +1,7 @@
+@props([
+    'rowCount' => 10,
+])
+
 <x-table-skeleton
     device="desktop"
     :items="[
@@ -23,5 +27,6 @@
         ],
     ]"
     :component-properties="['rounded' => false]"
+    :row-count="$rowCount"
     encapsulated
 />
