+++ resources/views/components/tables/headers/mobile/encapsulated/address.blade.php
@@ -1,0 +1,8 @@
+<div class="flex justify-center items-center space-x-2">
+    <x-general.identity
+        :model="$model"
+        without-username
+        without-reverse
+        without-icon
+    />
+</div>
