+++ resources/views/components/tables/headers/mobile/encapsulated/rank.blade.php
@@ -1,0 +1,8 @@
+@props([
+    'index',
+    'results',
+])
+
+<div class="font-semibold dark:text-theme-secondary-200">
+    {{ $index + ($results->currentPage() - 1) * $results->perPage()  }}
+</div>
