+++ resources/views/components/tables/mobile/skeleton/wallet-voters.blade.php
@@ -1,0 +1,4 @@
+<x-table-skeleton
+    device="mobile"
+    :items="['address', 'number', 'number']"
+/>
