+++ resources/views/components/tables/rows/desktop/encapsulated/address.blade.php
@@ -2,6 +2,7 @@
     'model',
     'withoutTruncate' => false,
     'withoutUsername' => false,
+    'withoutClipboard' => false,
 ])
 
 <span class="flex justify-between w-full text-sm leading-[17px]">
@@ -13,20 +14,28 @@
             without-icon
         >
             <x-slot name="address">
-                <span class="xl:hidden">
-                    <x-truncate-middle>{{ $model->address() }}</x-truncate-middle>
-                </span>
-                <span class="hidden xl:inline">
-                    {{ $model->address() }}
-                </span>
+                @unless ($withoutTruncate)
+                    <span class="xl:hidden">
+                        <x-truncate-middle>{{ $model->address() }}</x-truncate-middle>
+                    </span>
+                    <span class="hidden xl:inline">
+                        {{ $model->address() }}
+                    </span>
+                @else
+                    <span class="inline">
+                        {{ $model->address() }}
+                    </span>
+                @endif
             </x-slot>
         </x-general.identity>
     </span>
 
-    <x-ark-clipboard
-        :value="$model->address()"
-        :tooltip-content="trans('pages.wallet.address_copied')"
-        class="mr-3 transition text-theme-primary-400 dark:text-theme-secondary-600 hover:text-theme-primary-700"
-        no-styling
-    />
+    @unless ($withoutClipboard)
+        <x-ark-clipboard
+            :value="$model->address()"
+            :tooltip-content="trans('pages.wallet.address_copied')"
+            class="mr-3 transition text-theme-primary-400 dark:text-theme-secondary-600 hover:text-theme-primary-700"
+            no-styling
+        />
+    @endunless
 </span>
