+++ resources/views/components/tables/rows/mobile.blade.php
@@ -1,4 +1,4 @@
-<div class="rounded border border-theme-secondary-300 dark:border-theme-secondary-800">
+<div class="text-sm rounded border border-theme-secondary-300 dark:border-theme-secondary-800">
     <div class="flex justify-between items-center py-3 px-4 rounded-t dark:rounded-t-sm bg-theme-secondary-100 dark:bg-theme-secondary-800">
         {{ $header }}
     </div>
