+++ resources/views/components/tables/rows/mobile/encapsulated/balance.blade.php
@@ -1,9 +1,11 @@
-<div>
-    <span class="font-semibold">
-        @lang('labels.balance')
-    </span>
+@props(['model'])
 
-    <span class="font-semibold text-theme-secondary-900 dark:text-theme-secondary-200">
-        <x-general.amount-fiat-tooltip :amount="$model->balance()" :fiat="$model->balanceFiat()" />
-    </span>
+<div {{ $attributes->class('space-y-2 sm:flex sm:flex-col sm:justify-center') }}>
+    <div class="text-sm font-semibold leading-[17px] dark:text-theme-secondary-500">
+        @lang('tables.wallets.balance_currency', ['currency' => Network::currency()])
+    </div>
+
+    <div class="inline-block font-semibold text-theme-secondary-900 dark:text-theme-secondary-50">
+        <x-tables.rows.desktop.encapsulated.balance :model="$model" />
+    </div>
 </div>
