+++ resources/views/components/tables/rows/mobile/encapsulated/username.blade.php
@@ -1,7 +1,17 @@
-<div>
-    <span class="font-semibold">
-        @lang('labels.name')
-    </span>
+@props(['model'])
 
-    <span class="font-semibold">{{ $model->username() }}</span>
+<div {{ $attributes->class('space-y-2 sm:flex sm:flex-col sm:justify-center font-semibold') }}>
+    <div class="text-sm leading-[17px] dark:text-theme-secondary-500">
+        @lang('labels.name')
+    </div>
+
+    @if ($model->username())
+        <div class="inline-block text-theme-secondary-900 dark:text-theme-secondary-50">
+            {{ $model->username() }}
+        </div>
+    @else
+        <div class="text-theme-secondary-500 dark:text-theme-secondary-700">
+            @lang('general.na')
+        </div>
+    @endif
 </div>
