+++ resources/views/components/wallet/tables/transactions.blade.php
@@ -1,0 +1,9 @@
+@props(['wallet'])
+
+<div
+    x-show="tab === 'transactions'"
+    id="transactions-list"
+    class="w-full"
+>
+    <livewire:wallet-transaction-table :wallet="$wallet" />
+</div>
