+++ resources/views/components/wallet/tables/voters.blade.php
@@ -1,0 +1,9 @@
+@props(['wallet'])
+
+<div
+    x-show="tab === 'voters'"
+    id="voters-list"
+    class="w-full"
+>
+    <livewire:wallet-voter-table :wallet="$wallet" />
+</div>
