+++ tests/Feature/Http/Livewire/TopAccountsTableTest.php
@@ -7,7 +7,6 @@ use App\Http\Livewire\TopAccountsTable;
 use App\Models\Scopes\OrderByBalanceScope;
 use App\Models\Wallet;
 use App\Services\Cache\NetworkCache;
-use App\Services\NumberFormatter;
 use App\ViewModels\ViewModelFactory;
 use Livewire\Livewire;
 
@@ -20,7 +19,10 @@ it('should list the first page of records', function () {
 
     foreach (ViewModelFactory::paginate(Wallet::withScope(OrderByBalanceScope::class)->paginate())->items() as $wallet) {
         $component->assertSee($wallet->address());
-        $component->assertSee(NumberFormatter::currency($wallet->balance(), Network::currency()));
+        $component->assertSeeInOrder([
+            Network::currency(),
+            $wallet->balance(),
+        ]);
     }
 });
